# openapi.yaml
openapi: 3.1.0
info:
  title: fleet-rlm
  version: 0.5.0
paths:
...
  /api/v1/sessions/state:
    get:
      tags:
      - sessions
      summary: List Session State
      description: Return lightweight summaries of active/restored in-memory session
        state.
      operationId: list_session_state_api_v1_sessions_state_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStateResponse'
        '401':
          description: Authentication is required or the provided token is invalid.
        '503':
          description: Session state is unavailable because server startup is incomplete.
      security:
      - HTTPBearer: []
  /api/v1/sessions:
    get:
      tags:
      - sessions
      summary: List session history
      description: Paginated list of durable session transcripts with search and status
        filters.
      operationId: list_sessions_endpoint_api_v1_sessions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Full-text search on title
          title: Search
        description: Full-text search on title
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by status (active, archived)
          title: Status
        description: Filter by status (active, archived)
      - name: created_after
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Filter sessions created on or after this date (ISO 8601)
          title: Created After
        description: Filter sessions created on or after this date (ISO 8601)
      - name: created_before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Filter sessions created on or before this date (ISO 8601)
          title: Created Before
        description: Filter sessions created on or before this date (ISO 8601)
      - name: model_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by exact model name
          title: Model Name
        description: Filter by exact model name
      - name: model_provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by exact model provider
          title: Model Provider
        description: Filter by exact model provider
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Page size
          default: 20
          title: Limit
        description: Page size
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          description: Authentication is required or the provided token is invalid.
        '403':
          description: The caller does not have permission to access this resource.
        '503':
          description: Session services are unavailable because server startup is
            incomplete.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sessions/{session_id}:
    get:
      tags:
      - sessions
      summary: Get session detail
      description: Return session metadata and turn count for a specific session.
      operationId: get_session_detail_api_v1_sessions__session_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          description: Identifier of the session to inspect.
          title: Session Id
        description: Identifier of the session to inspect.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetailResponse'
        '401':
          description: Authentication is required or the provided token is invalid.
        '403':
          description: The caller does not have permission to access this resource.
        '503':
          description: Session services are unavailable because server startup is
            incomplete.
        '404':
          description: Session not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - sessions
      summary: Patch session metadata
      description: Update session title and/or metadata_json. Returns the updated
        session snapshot.
      operationId: patch_session_endpoint_api_v1_sessions__session_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          description: Identifier of the session to update.
          title: Session Id
        description: Identifier of the session to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetailResponse'
        '401':
          description: Authentication is required or the provided token is invalid.
        '403':
          description: The caller does not have permission to access this resource.
        '503':
          description: Session services are unavailable because server startup is
            incomplete.
        '404':
          description: Session not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sessions
      summary: Archive session
      description: Soft-delete (archive) a session. Returns success when archived,
        404 if not found or not owned.
      operationId: delete_session_endpoint_api_v1_sessions__session_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          description: Identifier of the session to archive.
          title: Session Id
        description: Identifier of the session to archive.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDeleteResponse'
        '401':
          description: Authentication is required or the provided token is invalid.
        '403':
          description: The caller does not have permission to access this resource.
        '503':
          description: Session services are unavailable because server startup is
            incomplete.
        '404':
          description: Session not found.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/runtime/status:
    get:
      tags:
      - runtime
      summary: Get Runtime Status
      description: Return the combined runtime readiness, model, and provider diagnostics
        snapshot.
      operationId: get_runtime_status_api_v1_runtime_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeStatusResponse'
        '401':
          description: Authentication is required or the provided token is invalid.
        '503':
          description: Runtime services are unavailable because server startup is
            incomplete.
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
          description: Structured list of request validation issues returned by FastAPI.
      type: object
      title: HTTPValidationError
...
    RuntimeStatusResponse:
      properties:
        app_env:
          type: string
          title: App Env
          description: Current application environment, such as `local` or `prod`.
        write_enabled:
          type: boolean
          title: Write Enabled
          description: Whether runtime settings writes are currently allowed.
        ready:
          type: boolean
          title: Ready
          description: Whether critical runtime services are ready to serve requests.
        active_models:
          $ref: '#/components/schemas/RuntimeActiveModels'
          description: Resolved planner and delegate model identities.
        sandbox_provider:
          type: string
          const: daytona
          title: Sandbox Provider
          description: Active sandbox backend selected for runtime execution and volume
            browsing.
          default: daytona
        llm:
          additionalProperties: true
          type: object
          title: Llm
          description: Language-model configuration and readiness diagnostics.
        mlflow:
          additionalProperties: true
          type: object
          title: Mlflow
          description: MLflow enablement and startup diagnostics.
        daytona:
          additionalProperties: true
          type: object
          title: Daytona
          description: Daytona configuration and readiness diagnostics.
        tests:
          $ref: '#/components/schemas/RuntimeTestCache'
          description: Cached runtime connectivity test results exposed in the Settings
            UI.
        guidance:
          items:
            type: string
          type: array
          title: Guidance
          description: Human-readable remediation steps for incomplete runtime setup.
      type: object
      required:
      - app_env
      - write_enabled
      - ready
      - active_models
      - tests
      title: RuntimeStatusResponse
      description: Combined readiness and diagnostics snapshot for the runtime settings
        UI.
...
    SessionDeleteResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          description: Whether the session was archived successfully.
          default: true
      type: object
      title: SessionDeleteResponse
      description: Result payload after archiving a session.
    SessionDetailResponse:
      properties:
        id:
          type: string
          title: Id
          description: Durable session identifier.
        title:
          type: string
          title: Title
          description: Human-readable session title.
        status:
          type: string
          title: Status
          description: Session status (active, archived).
        model_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Name
          description: Model used in session.
        external_session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Session Id
          description: Canonical runtime session identifier.
        workspace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Workspace Id
          description: Workspace context.
        turn_count:
          type: integer
          title: Turn Count
          description: Total number of turns in this session.
        created_at:
          type: string
          title: Created At
          description: ISO-8601 creation timestamp.
        updated_at:
          type: string
          title: Updated At
          description: ISO-8601 last-update timestamp.
      type: object
      required:
      - id
      - title
      - status
      - turn_count
      - created_at
      - updated_at
      title: SessionDetailResponse
      description: Full session detail with turn count.
...
    SessionListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SessionListItem'
          type: array
          title: Items
          description: Session list items.
        total:
          type: integer
          title: Total
          description: Total matching sessions.
        offset:
          type: integer
          title: Offset
          description: Current pagination offset.
        limit:
          type: integer
          title: Limit
          description: Current page size.
        has_more:
          type: boolean
          title: Has More
          description: Whether more results exist beyond this page.
      type: object
      required:
      - items
      - total
      - offset
      - limit
      - has_more
      title: SessionListResponse
      description: Paginated session list.
    SessionPatchRequest:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          description: New human-readable session title.
        metadata_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata Json
          description: New metadata dictionary to merge or replace session metadata.
      type: object
      title: SessionPatchRequest
      description: Patch body for updating session metadata.
...
    SessionStateResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          description: Whether the session-state query completed successfully.
          default: true
        sessions:
          items:
            $ref: '#/components/schemas/SessionStateSummary'
          type: array
          title: Sessions
          description: Active or restored session summaries currently known to the
            server.
      type: object
      title: SessionStateResponse
      description: Response body for the session-state summary endpoint.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
          description: Location path identifying where the validation error occurred.
        msg:
          type: string
          title: Message
          description: Human-readable validation failure message.
        type:
          type: string
          title: Error Type
          description: Pydantic validation error type identifier.
        input:
          title: Input
          description: Input value that failed validation, when available.
        ctx:
          type: object
          title: Context
          description: Optional structured validation context for templated error
            messages.
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Bearer token used when HTTP authentication is enabled. When auth
        is optional, requests without a token fall back to the configured default
        server identity.
      scheme: bearer
      bearerFormat: JWT